requirements.txt
playwright>=1.44.0
google-genai
python-telegram-bot>=21.3
python-dotenv==1.0.1
apscheduler>=3.10.0
